+++ src/arms_gibbs.cpp
@@ -2,6 +2,84 @@
 
 #include "arms_gibbs.h"
 
+// Multivariate ARMS via Gibbs sampler for logistic regression
+void ARMS_Gibbs::arms_gibbs_beta_logistic(
+    const armsParmClass& armsPar,
+    const hyperparClass& hyperpar,
+    arma::mat& currentPars,
+    arma::umat gammas,
+    double tau0Sq,
+    double tauSq,
+    const DataClass &dataclass
+)
+{
+    // dimensions
+    unsigned int N = dataclass.X.n_rows;
+    unsigned int p = dataclass.X.n_cols;
+
+    // objects for arms()
+    double minD = armsPar.betaMin;
+    double maxD = armsPar.betaMax;
+
+    std::vector<double> xinit(armsPar.ninit); // Use std::vector instead of VLA
+
+    arma::vec xinit0 = arma::linspace( minD+1.0e-10, maxD-1.0e-10, armsPar.ninit );
+    for (unsigned int i = 0; i < armsPar.ninit; ++i)
+        xinit[i] = xinit0[i];
+
+    dataS *mydata = (dataS *)malloc(sizeof (dataS));
+
+    // gammas = arma::join_cols(arma::ones<arma::urowvec>(1), gammas);
+    currentPars.elem(arma::find(gammas == 0)).fill(0.);
+    mydata->currentPars = currentPars.memptr();
+    mydata->p = p;
+    mydata->N = N;
+    mydata->tau0Sq = tau0Sq;
+    mydata->tauSq = tauSq;
+    // mydata->mu = mu.memptr();
+    mydata->X = dataclass.X.memptr();
+    mydata->y = dataclass.y.memptr();
+    mydata->event = dataclass.event.memptr();
+
+    // Gibbs sampling
+
+    for (unsigned int j = 0; j < p; ++j)
+    {
+        if (gammas(j))
+        {
+            mydata->jj = j;
+
+            double xprev = currentPars(j);
+            std::vector<double> xsamp(armsPar.nsamp);
+
+            double qcent[1], xcent[1];
+            int neval, ncent = 0;
+
+            int err;
+            double convex = armsPar.convex;
+            err = ARMS::arms (
+                      xinit.data(), armsPar.ninit, &minD, &maxD,
+                      EvalFunction::log_dens_betas_logistic, mydata,
+                      &convex, armsPar.npoint,
+                      armsPar.metropolis, &xprev, xsamp.data(),
+                      armsPar.nsamp, qcent, xcent, ncent, &neval);
+
+            // check ARMS validity
+            if (err > 0)
+                Rprintf("In arms_gibbs_beta(): error code in ARMS = %d.\n", err);
+            if (std::isnan(xsamp[armsPar.nsamp-1]))
+                Rprintf("In arms_gibbs_beta(): NaN generated, possibly due to overflow in (log-)density (e.g. with densities involving exp(exp(...))).\n");
+            if (xsamp[armsPar.nsamp-1] < minD || xsamp[armsPar.nsamp-1] > maxD)
+                Rprintf("In arms_gibbs_beta(): %d-th sample out of range [%f, %f] (fused domain). Got %f.\n", armsPar.nsamp, minD, maxD, xsamp[armsPar.nsamp-1]);
+
+            currentPars(j) = xsamp[armsPar.nsamp - 1];
+
+        }
+    }
+
+    free(mydata);
+}
+
 // Multivariate ARMS via Gibbs sampler for all responses of Dirichlet model
 void ARMS_Gibbs::arms_gibbs_beta_dirichlet(
     const armsParmClass& armsPar,