+++ src/drive.cpp
@@ -2,6 +2,7 @@
 
 
 #include "BVS_gaussian.h"
+#include "BVS_logistic.h"
 #include "BVS_probit.h"
 #include "BVS_weibull.h"
 #include "BVS_dirichlet.h"
@@ -337,7 +338,28 @@ Rcpp::List run_mcmc(
         break;
 
     case Family_Type::logit :
-        ::Rf_error("Not yet implemented!");
+        BVS_logistic::mcmc(
+            nIter,
+            burnin,
+            thin,
+            tau0Sq,
+            tauSq,
+            betas,
+            gammas,
+            gammaProposal,
+            gammaSampler,
+            armsPar,
+            hyperpar,
+            dataclass,
+
+            beta_mcmc,
+            beta_post,
+            gamma_mcmc,
+            gamma_post,
+            gamma_acc_count,
+            loglikelihood_mcmc,
+            tauSq_mcmc
+        );
         break;
 
     case Family_Type::probit :